Add command line option to debug process
This adds the command line argument --debug.
All command line args after --debug won't be interpreted for dnSpy.
The first command line arg after --debug is interpreted as the path to
the application to debug and the rest of the command line args will be
forwarded to the application to debug. This is necessary if one wants to
use the windows function to automatically start a debugger if an
application is started, because the application and it's command line args
are passed as the last args to the debugger.